Struct leo_ast::common::identifier::Identifier

pub struct Identifier {
    pub name: Symbol,
    pub span: Span,
}

An identifier in a program.

Attention - When adding or removing fields from this struct, please remember to update its Serialize and Deserialize implementation to reflect the new struct instantiation.

Fields
name: Symbol

The symbol that the user wrote, e.g., foo.

span: Span

A span locating where the identifier occurred in the source.

Implementations

impl Identifier

pub fn new(name: Symbol) -> Self

Constructs a new identifier with name and a default span.

pub fn matches(&self, other: &Self) -> bool

Check if the Identifier name matches the other name.