leo_ast::common

Type Alias ImportGraph

pub type ImportGraph = DiGraph<Symbol>;

An import dependency graph.

Aliased Type

pub struct ImportGraph {
    nodes: IndexSet<Rc<Symbol>>,
    edges: IndexMap<Rc<Symbol>, IndexSet<Rc<Symbol>>>,
}

Fields
nodes: IndexSet<Rc<Symbol>>

The set of nodes in the graph.

edges: IndexMap<Rc<Symbol>, IndexSet<Rc<Symbol>>>

The directed edges in the graph. Each entry in the map is a node in the graph, and the set of nodes that it points to.