leo_ast::common

Type Alias StructGraph

pub type StructGraph = DiGraph<Vec<Symbol>>;

A struct dependency graph. The Vec<Symbol> is to the absolute path to each struct

Aliased Type

pub struct StructGraph {
    nodes: IndexSet<Rc<Vec<Symbol>>>,
    edges: IndexMap<Rc<Vec<Symbol>>, IndexSet<Rc<Vec<Symbol>>>>,
}

Fields
nodes: IndexSet<Rc<Vec<Symbol>>>

The set of nodes in the graph.

edges: IndexMap<Rc<Vec<Symbol>>, IndexSet<Rc<Vec<Symbol>>>>

The directed edges in the graph. Each entry in the map is a node in the graph, and the set of nodes that it points to.