leo_ast

Module constructor

Modules

admin 🔒

checksum 🔒

noupgrade 🔒

Structs

Constructor

A constructor definition.

Enums

UpgradeVariant

The upgrade variant.

Functions

get_type_from_snarkvm_literal 🔒

snarkvm_admin_constructor

This is the expected snarkVM constructor bytecode for a program that is only upgradable by a fixed admin.

snarkvm_checksum_constructor

This is the required snarkVM constructor bytecode for a program that is only upgradable if the new program’s checksum matches the one declared in a pre-determined mapping.

snarkvm_noupgrade_constructor

This is the expected snarkVM constructor bytecode for a program that is not upgradable.