Struct leo_ast::input::input::Input

pub struct Input {
    name: String,
    program_input: ProgramInput,
    program_state: ProgramState,
}

Fields
name: Stringprogram_input: ProgramInputprogram_state: ProgramState
Implementations

impl Input

pub fn new() -> Self

pub fn empty(&self) -> Self

Returns an empty version of this struct with None values. Called during constraint synthesis to provide private input variables.

pub fn len(&self) -> usize

Returns the number of input variables to pass into the main program function

pub fn set_main_input(&mut self, input: MainInput)

Manually set the input variables to the main program function

pub fn parse_input(&mut self, file: File<'_>) -> Result<(), InputParserError>

Parse all input variables included in a file and store them in self.

pub fn parse_state(&mut self, file: File<'_>) -> Result<(), InputParserError>

Parse all state variables included in a file and store them in self.

pub fn get(&self, name: &str) -> Option<Option<InputValue>>

Returns the main function input value with the given name.

pub fn get_constant(&self, name: &str) -> Option<Option<InputValue>>

Returns the constant input value with the given name.

pub fn get_registers(&self) -> &Registers

Returns the runtime register input values

pub fn get_record(&self) -> &Record

Returns the runtime record input values

pub fn get_state(&self) -> &State

Returns the runtime state input values

pub fn get_state_leaf(&self) -> &StateLeaf

Returns the runtime state leaf input values