Function leo_ast::input::input_value::parse_array_dimensions

fn parse_array_dimensions(
    array_dimensions_type: ArrayDimensions, 
    span: &Span<'_>
) -> Result<Vec<usize>, InputParserError>

Returns a new vector of usize values from an ArrayDimensions type.

Attempts to parse each dimension in the array from a String to a usize value. If parsing is successful, the usize value is appended to the return vector. If parsing fails, an error is returned.