leo_ast::interpreter_value

Enum Value

pub enum Value {

    Unit,
    Bool(bool),
    U8(u8),
    U16(u16),
    U32(u32),
    U64(u64),
    U128(u128),
    I8(i8),
    I16(i16),
    I32(i32),
    I64(i64),
    I128(i128),
    Group(SvmGroup),
    Field(SvmField),
    Scalar(SvmScalar),
    Array(Vec<Value>),
    Repeat(Box<Value>, Box<Value>),
    Tuple(Vec<Value>),
    Address(SvmAddress),
    Future(Future),
    Struct(StructContents),
    Unsuffixed(String),

}

A Leo value of any type.

Mappings and functions aren’t considered values.

Variants

Unit

Bool(bool)

U8(u8)

U16(u16)

U32(u32)

U64(u64)

U128(u128)

I8(i8)

I16(i16)

I32(i32)

I64(i64)

I128(i128)

Group(SvmGroup)

Field(SvmField)

Scalar(SvmScalar)

Array(Vec<Value>)

Repeat(Box<Value>, Box<Value>)

Tuple(Vec<Value>)

Address(SvmAddress)

Future(Future)

Struct(StructContents)

Unsuffixed(String)

Implementations

impl Value

pub fn to_fields(&self) -> Vec<SvmField>

pub fn gte(&self, rhs: &Self) -> Result<bool>

pub fn lte(&self, rhs: &Self) -> Result<bool>

pub fn lt(&self, rhs: &Self) -> Result<bool>

pub fn gt(&self, rhs: &Self) -> Result<bool>

pub fn neq(&self, rhs: &Self) -> Result<bool>

pub fn eq(&self, rhs: &Self) -> Result<bool>

Are the values equal, according to SnarkVM?

We use this rather than the Eq trait so we can fail when comparing values of different types, rather than just returning false.

pub fn inc_wrapping(&self) -> Self

pub fn generator() -> Self

Return the group generator.

pub fn simple_shl(&self, shift: u32) -> Self

Doesn’t correspond to Aleo’s shl, because it does not fail when set bits are shifted out.

pub fn simple_shr(&self, shift: u32) -> Self

pub fn cast(&self, cast_type: &Type) -> Option<Value>

Convert to the given type if possible under Aleo casting rules.

pub fn resolve_if_unsuffixed( &self, ty: &Option<Type>, span: Span, ) -> Result<Value>

Resolves an unsuffixed literal to a typed Value using the provided optional Type. If the value is unsuffixed and a type is provided, parses the string into the corresponding Value variant. Handles integers of various widths and special types like Field, Group, and Scalar. If no type is given or the value is already typed, returns the original value. Returns an error if type inference is not possible or parsing fails.

impl Value

pub fn get_numeric_type(&self) -> Option<Type>

Gets the type of a Value but only if it is an integer, a field, a group, or a scalar. Return None otherwise. These are the only types that an unsuffixed literal can have.