leo_ast::interpreter_value

Module value

Macros

impl_from_integer ğŸ”’

impl_from_literal ğŸ”’

Structs

StructContents

SvmIdentifierParam ğŸ”’

An identifier is an immutable UTF-8 string, represented as a constant field element in the CurrentNetwork.

SvmSignature ğŸ”’

TestnetV0 ğŸ”’

Value

Enums

AsyncExecution

Plaintext ğŸ”’

SvmLiteralParam ğŸ”’

The literal enum represents all supported types in snarkVM.

SvmValueParam ğŸ”’

ValueVariants ğŸ”’

Traits

TryAsRef

Functions

hash_plaintext ğŸ”’

plaintext_to_expression ğŸ”’

Type Aliases

Address ğŸ”’

Boolean ğŸ”’

CurrentNetwork ğŸ”’

Field ğŸ”’

Future ğŸ”’

Group ğŸ”’

ProgramID ğŸ”’

Scalar ğŸ”’

Signature ğŸ”’

SvmIdentifier ğŸ”’

SvmLiteral ğŸ”’

SvmPlaintext ğŸ”’

SvmValue ğŸ”’