leo_ast::interpreter_value::value

Enum SvmLiteralParam

pub(crate) enum SvmLiteralParam<N>
where
    N: Network,
{

    Address(Address<N>),
    Boolean(Boolean<N>),
    Field(Field<N>),
    Group(Group<N>),
    I8(Integer<N, i8>),
    I16(Integer<N, i16>),
    I32(Integer<N, i32>),
    I64(Integer<N, i64>),
    I128(Integer<N, i128>),
    U8(Integer<N, u8>),
    U16(Integer<N, u16>),
    U32(Integer<N, u32>),
    U64(Integer<N, u64>),
    U128(Integer<N, u128>),
    Scalar(Scalar<N>),
    Signature(Box<Signature<N>>),
    String(StringType<N>),

}

The literal enum represents all supported types in snarkVM.

Variants

Address(Address<N>)

The Aleo address type.

Boolean(Boolean<N>)

The boolean type.

Field(Field<N>)

The field type.

Group(Group<N>)

The group type.

I8(Integer<N, i8>)

The 8-bit signed integer type.

I16(Integer<N, i16>)

The 16-bit signed integer type.

I32(Integer<N, i32>)

The 32-bit signed integer type.

I64(Integer<N, i64>)

The 64-bit signed integer type.

I128(Integer<N, i128>)

The 128-bit signed integer type.

U8(Integer<N, u8>)

The 8-bit unsigned integer type.

U16(Integer<N, u16>)

The 16-bit unsigned integer type.

U32(Integer<N, u32>)

The 32-bit unsigned integer type.

U64(Integer<N, u64>)

The 64-bit unsigned integer type.

U128(Integer<N, u128>)

The 128-bit unsigned integer type.

Scalar(Scalar<N>)

The scalar type.

Signature(Box<Signature<N>>)

The signature type.

String(StringType<N>)

The string type.

Implementations

impl<N> Literal<N>
where N: Network,

pub fn cast(&self, to_type: LiteralType) -> Result<Literal<N>, Error>

Casts the literal to the given literal type.

This method checks that the cast does not lose any bits of information, and returns an error if it does.

The hierarchy of casting is as follows:

(Address, Group) <-> Field <-> Scalar <-> Integer <-> Boolean

Signature (not supported)

String (not supported)

Note that casting to left along the hierarchy always preserves information.

impl<N> Literal<N>
where N: Network,

pub fn cast_lossy(&self, to_type: LiteralType) -> Result<Literal<N>, Error>

Casts the literal to the given literal type, with lossy truncation.

This method makes a best-effort attempt to preserve all bits of information, but it is not guaranteed to do so.

The hierarchy of casting is as follows:

(Address, Group) <-> Field <-> Scalar <-> Integer <-> Boolean

Signature (not supported)

String (not supported)

Note that casting to left along the hierarchy always preserves information.

impl<N> Literal<N>
where N: Network,

pub fn from_bits_le(variant: u8, bits_le: &[bool]) -> Result<Literal<N>, Error>

Initializes a new literal from a list of little-endian bits without trailing zeros.

pub fn from_bits_be(variant: u8, bits_be: &[bool]) -> Result<Literal<N>, Error>

Initializes a new literal from a list of big-endian bits without leading zeros.

impl<N> Literal<N>
where N: Network,

pub fn sample<R>(literal_type: LiteralType, rng: &mut R) -> Literal<N>
where R: Rng + CryptoRng,

Returns a randomly-sampled literal of the given literal type.

impl<N> Literal<N>
where N: Network,

pub fn size_in_bits(&self) -> u16

Returns the number of bits of this literal.

impl<N> Literal<N>
where N: Network,

pub fn size_in_bytes(&self) -> u16

Returns the size in bytes of this literal.

impl<N> Literal<N>
where N: Network,

pub fn to_type(&self) -> LiteralType

Returns the type name of the literal.

impl<N> Literal<N>
where N: Network,

pub fn variant(&self) -> u8

Returns the variant of the literal.