leo_ast::interpreter_value::value

Struct Value

pub struct Value {
    pub id: Option<Location>,
    pub(crate) contents: ValueVariants,
}

Fields
id: Option<Location>contents: ValueVariants
Implementations

impl Value

pub fn eq(&self, rhs: &Self) -> Result<bool>

Are the values equal, according to SnarkVM?

We use this rather than the Eq trait so we can fail when comparing values of different types, rather than just returning false.

pub fn resolve_if_unsuffixed( &self, ty: &Option<Type>, span: Span, ) -> Result<Value>

Resolves an unsuffixed literal to a typed Value using the provided optional Type. If the value is unsuffixed and a type is provided, parses the string into the corresponding Value variant. Handles integers of various widths and special types like Field, Group, and Scalar. If no type is given or the value is already typed, returns the original value. Returns an error if type inference is not possible or parsing fails.

impl Value

pub fn lt(&self, rhs: &Self) -> Option<bool>

pub fn lte(&self, rhs: &Self) -> Option<bool>

pub fn gt(&self, rhs: &Self) -> Option<bool>

pub fn gte(&self, rhs: &Self) -> Option<bool>

pub fn inc_wrapping(&self) -> Option<Self>

pub fn add(&self, i: usize) -> Option<Self>

pub fn cast(&self, ty: &Type) -> Option<Self>

pub fn cast_lossy(&self, ty: &LiteralType) -> Option<Self>

pub fn generator() -> Self

Return the group generator.

pub fn as_u32(&self) -> Option<u32>

pub fn as_i128(&self) -> Option<i128>

pub fn array_index(&self, i: usize) -> Option<Self>

pub fn array_index_set(&mut self, i: usize, value: Self) -> Option<()>

pub fn tuple_len(&self) -> Option<usize>

pub fn tuple_index(&self, i: usize) -> Option<Self>

pub fn tuple_index_set(&mut self, i: usize, value: Self) -> Option<()>

pub fn as_future(&self) -> Option<&[AsyncExecution]>

pub fn accesses( &self, accesses: impl IntoIterator<Item = Access<TestnetV0>>, ) -> Option<Value>

fn accesses_impl( &self, accesses: &mut dyn Iterator<Item = Access<TestnetV0>>, ) -> Option<Value>

pub fn member_access(&self, member: Symbol) -> Option<Self>

pub fn member_set(&mut self, member: Symbol, value: Value) -> Option<()>

pub fn make_future( program_name: Symbol, function: Symbol, arguments: impl IntoIterator<Item = Value>, ) -> Option<Value>

fn make_future_impl( program_name: Symbol, function: Symbol, arguments: &mut dyn Iterator<Item = Value>, ) -> Option<Value>

pub fn make_unit() -> Self

pub fn make_struct( contents: impl Iterator<Item = (Symbol, Value)>, program: Symbol, path: Vec<Symbol>, ) -> Self

pub fn make_record( contents: impl Iterator<Item = (Symbol, Value)>, program: Symbol, path: Vec<Symbol>, ) -> Self

pub fn make_array(contents: impl Iterator<Item = Value>) -> Self

pub fn make_tuple(contents: impl IntoIterator<Item = Value>) -> Self

pub fn make_string(s: String) -> Self

pub fn get_numeric_type(&self) -> Option<Type>

Gets the type of a Value but only if it is an integer, a field, a group, or a scalar. Return None otherwise. These are the only types that an unsuffixed literal can have.

pub fn to_expression( &self, span: Span, node_builder: &NodeBuilder, ty: &Type, struct_lookup: &dyn Fn(&[Symbol]) -> Vec<(Symbol, Type)>, ) -> Option<Expression>