leo_ast::passes

Module consumer

This module contains a Consumer trait for the AST. Consumers are used to completely transform the AST without any restrictions on the output.

Traits

ConstructorConsumer

A Consumer trait for constructors in the AST.

ExpressionConsumer

A Consumer trait for expressions in the AST.

FunctionConsumer

A Consumer trait for functions in the AST.

ImportConsumer

A Consumer trait for imported programs in the AST.

MappingConsumer

A Consumer trait for mappings in the AST.

ModuleConsumer

A Consumer trait for modules in the AST.

ProgramConsumer

A Consumer trait for the program represented by the AST.

ProgramScopeConsumer

A Consumer trait for program scopes in the AST.

StatementConsumer

A Consumer trait for statements in the AST.

StructConsumer

A Consumer trait for structs in the AST.