Module leo_ast::passes::reconstructor

This module contains a Reconstructor trait for the AST. It implements default methods for each node to be made given the information of the old node.

Traits

ExpressionReconstructor

A Reconstructor trait for expressions in the AST.

ProgramReconstructor

A Reconstructor trait for the program represented by the AST.

StatementReconstructor

A Reconstructor trait for statements in the AST.