Struct leo_ast::reducer::Canonicalizer

pub struct Canonicalizer {
    circuit_name: Option<Identifier>,
    in_circuit: bool,
}

Replace Self when it is in a enclosing circuit type. Error when Self is outside an enclosing circuit type. Tuple array types and expressions expand to nested arrays. Tuple array types and expressions error if a size of 0 is given. Compound operators become simple assignments. Functions missing output type return a empty tuple.

Fields
circuit_name: Option<Identifier>in_circuit: bool
Implementations

impl Canonicalizer

pub fn canonicalize_accesses(
&mut self,
start: Expression,
accesses: &[AssigneeAccess],
span: &Span
) -> Result<Box<Expression>>

pub fn compound_operation_converstion(
&mut self,
operation: &AssignOperation
) -> Result<BinaryOperation>

fn canonicalize_self_type(&self, type_option: Option<&Type>) -> Option<Type>

fn canonicalize_expression(&mut self, expression: &Expression) -> Expression

fn canonicalize_assignee_access(
&mut self,
access: &AssigneeAccess
) -> AssigneeAccess

fn canonicalize_assignee(&mut self, assignee: &Assignee) -> Assignee

fn canonicalize_block(&mut self, block: &Block) -> Block

fn canonicalize_statement(&mut self, statement: &Statement) -> Statement

fn canonicalize_circuit_member(
&mut self,
circuit_member: &CircuitMember
) -> CircuitMember