Crate leo_ast

The abstract syntax tree (ast) for a Leo program.

This module contains the Ast type, a wrapper around the Program type. The Ast type is intended to be parsed and modified by different passes of the Leo compiler. The Leo compiler can generate a set of R1CS constraints from any Ast.

Re-exports

pub use self::access::*;

pub use self::struct::*;

pub use self::common::*;

pub use self::expressions::*;

pub use self::functions::*;

pub use self::groups::*;

pub use self::input::*;

pub use self::mapping::*;

pub use self::passes::*;

pub use self::program::*;

pub use self::statement::*;

pub use self::types::*;

pub use self::value::*;

pub use common::node::*;

Modules

access

common

expressions

functions

groups

input

mapping

passes

This module contains both a Reducer and Visitor design pattern. These both iterate over the AST.

program

A Leo program consists of import statements and program scopes.

statement

struct

types

value

Macros

simple_node_impl

Structs

Ast

The abstract syntax tree (AST) for a Leo program.

Functions

normalize_json_value

Helper function to normalize AST JSON into a form compatible with tgc. This function will traverse the original JSON value and produce a new one under the following rules:

remove_key_from_json

Helper function to recursively filter keys from AST JSON