Module leo_ast::passes

This module contains both a Reducer and Visitor design pattern. These both iterate over the AST.

Re-exports

pub use consumer::*;

pub use reconstructor::*;

pub use visitor::*;

Modules

consumer

This module contains a Consumer trait for the AST. Consumers are used to completely transform the AST without any restrictions on the output.

reconstructor

This module contains a Reconstructor trait for the AST. It implements default methods for each node to be made given the information of the old node.

visitor

This module contains Visitor trait implementations for the AST. It implements default methods for each node to be made given the type of node its visiting.